Split up Kids.jsx
