+++ frontend/src/app/Kids/Kids.jsx
@@ -1,289 +1,15 @@
 import React from 'react';
-import {List, ListItem} from 'material-ui/List';
-import {Card, CardActions, CardHeader, CardMedia, CardTitle, CardText} from 'material-ui/Card';
-import Avatar from 'material-ui/Avatar';
-import FlatButton from 'material-ui/FlatButton';
-import TextField from 'material-ui/TextField';
-import ActionDone from 'material-ui/svg-icons/action/done';
-import ActionDelete from 'material-ui/svg-icons/action/delete';
-import ActionAnnouncement from 'material-ui/svg-icons/action/announcement';
 import ContentAdd from 'material-ui/svg-icons/content/add';
-import IconButton from 'material-ui/IconButton';
 import Snackbar from 'material-ui/Snackbar';
-import CircularProgress from 'material-ui/CircularProgress';
-import LinearProgress from 'material-ui/LinearProgress';
 import FloatingActionButton from 'material-ui/FloatingActionButton';
 
-import seedrandom from 'seedrandom';
-import tinycolor from 'tinycolor2';
-import FlipMove from 'react-flip-move';
 import ReactTimeout from 'react-timeout';
 import DocumentTitle from 'react-document-title';
-import Visibility from 'visibilityjs';
-import moment from 'moment';
 import _ from 'underscore';
 import $ from 'jquery';
 
 import styles from '../styles';
-import muiThemeable from 'material-ui/styles/muiThemeable';
-import {grey400, darkBlack, lightBlack} from 'material-ui/styles/colors';
-
-var material_palette = require("!json!./material_palette.json");
-
-var Announcement = muiThemeable()(React.createClass({
-    render: function() {
-	return (
-	    <ListItem primaryText={<strong>Announcement</strong>}
-		      secondaryText={<p><span style={{color: darkBlack}}>{this.props.message}</span></p>}
-		      leftAvatar={<Avatar icon={<ActionAnnouncement />}
-					  backgroundColor={this.props.muiTheme.palette.accent1Color} />}
-		      disabled={true}
-		      rightIconButton={this.props.instructor ?
-				       <IconButton onTouchTap={this.props.onRemove}><ActionDelete /> </IconButton> :
-				      null}
-	    />
-	);
-    }
-}));
-
-var KidsList = React.createClass({
-    checkMobile: function() {
-	if (window.matchMedia("only screen and (max-width: 760px)").matches) {
-            this.setState({s: styles.containerMobile});
-	}
-	else {
-	    this.setState({s: styles.container});
-	}
-    },
-    componentWillMount: function() {
-        this.checkMobile();
-    },
-    componentDidMount: function() {
-        window.addEventListener("resize", this.checkMobile);
-    },
-    componentWillUnmount: function() {
-        window.removeEventListener("resize", this.checkMobile);
-    },
-    render: function() {
-	var kidsNodes = this.props.data.map(function(kid) {
-	    return (
-		<Kid name={kid.name}
-		     room={kid.room}
-		     question={kid.question}
-		     id={kid.id}
-		     answer={kid.answer}
-		     timestamp={kid.timestamp}
-		     key={kid.id}
-		     onKidDelete={this.props.onKidDelete}
-		     onKidAnswer={this.props.onKidAnswer}
-		     username={this.props.username}
-		     instructor={this.props.instructor}
-		/>
-	    );
-	}.bind(this));
-	return (
-	    <div className="kidsList" style={this.state.s}>
-		<List style={styles.list}>
-		    {this.props.announcement ?
-		     <Announcement key={-41}
-				   message={this.props.announcement}
-				   onRemove={this.props.onRemoveAnnouncement}
-				   instructor={this.props.instructor}
-		     /> :
-		     null}
-		    <FlipMove enterAnimation={"elevator"}
-			      leaveAnimation={"elevator"}
-			      staggerDurationBy={10}
-			      staggerDelayBy={10}
-			      disableAllAnimations={Visibility.state() != "visible"}>
-			{kidsNodes}
-			<AddKid key={-42}
-				onKidSubmit={this.props.onKidSubmit}
-				editing={this.props.editing}
-				paused={this.props.paused}
-				adding={this.props.adding}
-				onAddExpandChange={this.props.onAddExpandChange}
-				onAddReduceChange={this.props.onAddReduceChange}
-			/>
-		    </FlipMove>
-		</List>
-	    </div>
-	);
-    }
-});
-
-var AddKid = React.createClass({
-    getInitialState: function() {
-	return {name: '',
-		room: '',
-		question:'',
-		attemptedSubmit: false
-	};
-    },
-    handleNameChange: function(e) {
-	this.setState({name: e.target.value});
-    },
-    handleRoomChange: function(e) {
-	this.setState({room: e.target.value});
-    },
-    handleQuestionChange: function(e) {
-	this.setState({question: e.target.value});
-    },
-    submitKid: function(e) {
-	this.props.onKidSubmit({
-	    name: this.state.name.trim(),
-	    room: this.state.room.trim(),
-	    question: this.state.question.trim(),
-	});
-	this.setState({room: '', question:'', attemptedSubmit: false});
-	this.props.onAddReduceChange();
-    },
-    handleKeyPress: function(target) {
-	if (target.charCode == 13) {
-	    if (this.state.name && this.state.room && this.state.question) {
-		this.submitKid();
-	    }
-	    else {
-		this.setState({attemptedSubmit: true});
-	    }
-	}
-    },
-    render: function() {
-	return (
-	    <Card expanded={this.props.adding} onExpandChange={this.props.onAddExpandChange}>
-		<CardHeader
-		    title={this.props.editing ? "Edit Question" : "New Question"}
-		    subtitle={this.props.paused ?
-			      "Submission currently disabled!" :
-			      "Expand to add!"}
-		    actAsExpander={true}
-		    showExpandableButton={true}
-		    avatar={<Avatar>{this.props.adding ? "-" : "+"}</Avatar>}
-		/>
-		<CardText expandable={true}>
-		    <TextField
-			hintText="Name"
-			floatingLabelText="Name"
-			errorText={!this.state.attemptedSubmit || this.state.name ?
-				   "" :
-				   "This field is required"}
-			value={this.state.name}
-			onChange={this.handleNameChange}
-			onKeyPress={this.handleKeyPress}
-			autoFocus={true}
-		    />
-		    <br/>
-		    <TextField
-			hintText="Room"
-			floatingLabelText="Room"
-			errorText={!this.state.attemptedSubmit || this.state.room ?
-				   "" :
-				   "This field is required"}
-			value={this.state.room}
-			onChange={this.handleRoomChange}
-			onKeyPress={this.handleKeyPress}
-		    />
-		    <br/>
-		    <TextField
-			hintText="Question"
-			floatingLabelText="Question"
-			errorText={!this.state.attemptedSubmit || this.state.question ?
-				   "" :
-				   "This field is required"}
-			value={this.state.question}
-			onChange={this.handleQuestionChange}
-			onKeyPress={this.handleKeyPress}
-		    />
-		    <br/>
-		    <CardActions>
-			<FlatButton
-			    label="Submit"
-			    disabled={!(this.state.name && this.state.room && this.state.question)}
-			    primary={true}
-			    onTouchTap={this.submitKid}
-			/>
-		    </CardActions>
-		</CardText>
-	    </Card>
-	);
-    }
-});
-
-var Kid = React.createClass({
-    getInitialState: function() {
-	var color = this.generateColor();
-	return ({color: color,
-		 complement: tinycolor(color).complement().toHexString(),
-		 timeDescription: "",
-		 timerId: null,
-	});
-    },
-    componentDidMount: function() {
-	this.setState({timeDescription: this.describeTime(this.props.timestamp),
-		       timerId: setTimeout(this.updateTime, 60000),
-	});
-    },
-    componentDidUpdate: function(prevProps, prevState) {
-	if (prevProps.timestamp != this.props.timestamp) {
-	    this.setState({timeDescription: this.describeTime(this.props.timestamp)});
-	}
-    },
-    componentWillUnmount: function() {
-	clearTimeout(this.state.timerId);
-    },
-    handleButtonTouchTap: function(e) {
-	this.props.onKidDelete({id: this.props.id});
-    },
-    handleTouchTap: function(e) {
-	this.props.onKidAnswer({id: this.props.id,
-				name: this.props.name,
-				room: this.props.room,
-				question: this.props.question,
-				answer: !this.props.answer,
-				});
-    },
-    generateColor: function() {
-	var n = Math.seedrandom(this.props.id);
-	var groups = Object.keys(material_palette);
-	var group = groups[Math.floor(groups.length * Math.random())];
-	var colors = Object.keys(material_palette[group]);
-	var coloridx = Math.floor(colors.length * Math.random());
-	var color = material_palette[group][colors[coloridx]];
-	return color;
-    },
-    describeTime: function(timestamp) {
-	return moment(timestamp).fromNow();
-    },
-    updateTime: function() {
-	/* var delta = 60000;
-	   if (moment.diff(this.state.timestamp, "hours") >= 1) {
-	   delta = 60000;
-	   }*/
-	this.setState({timerId: setTimeout(this.updateTime, 60000),
-		       timeDescription: this.describeTime(this.props.timestamp),
-	});
-    },
-    render: function() {
-	return (
-	    <Card>
-		<ListItem
-		    primaryText={this.props.name + " – " + this.props.room}
-		    secondaryText={<span>{this.props.question} <span style={styles.timestamp}> {this.state.timeDescription}</span></span>}
-		    leftAvatar={<Avatar backgroundColor={this.state.color} >{this.props.name[0]} </Avatar>}
-		    onTouchTap={this.props.instructor ? this.handleTouchTap : undefined}
-		    leftIcon={this.props.answer ? <CircularProgress color={this.state.complement} size={0.75} style={styles.progress}/> : null}
-		    rightIconButton={this.props.instructor || this.props.username == this.props.id ?
-		     (
-			 <IconButton
-			     onTouchTap={this.handleButtonTouchTap}>
-			     <ActionDone/>
-			 </IconButton>
-			) : null}
-		/>
-	    </Card>
-	);
-    }
-});
+import KidsList from './KidsList';
 
 var Kids = ReactTimeout(React.createClass({
     getInitialState: function() {
